refactor model list view action buttons
This changes the way the action buttons are being layered into the list
view. Instead of using the builtin actionButtons argument, the list
buttons are now rendered directly. This was done to enable more control
over the styling and functionality of the buttons. Also, a directive has
been added for the delete model functionality. The create query
directive has been modified to take a model as an input, in which case
the dropdown is only populated with that entry.